fix(background): let Stripe failures retry and make webhook redeliveries idempotent
- handleStripeWebhook no longer swallows rejections; failures reach the 500 path so Stripe redelivers
- deliveryId derives deterministically from the queue message id so every attempt of one message shares an id (receiver dedup contract)
- Stripe event dispatch collapses onto planForStripeEvent; signature helpers compose over capabilities' shared HMAC
- Split index.ts into webhook-consumer / stripe-endpoint / webhook-signing with index.ts as wiring only; one consumeBatch for both queue loops
